requirements.txt
ctransformers==0.2.27
faiss-cpu==1.7.4
gradio==4.31.2
llama_cpp_python==0.2.74
langchain==0.1.16
sentence-transformers==2.7.0
transformers~=4.40.0
torch~=2.2.2